Full data set rat epilepsy experiment

The data set entitled as „episet_full” contains data from five different studies. In these studies three different epilepsy models were investigated: a kindling model, an electrical and a chemical post-SE model. In all studies up to three different groups of animals were used. One group with implanted depth electrodes and generalized seizures (induced or spontaneous). The animals of the second group were implanted with depth electrodes and the third group was completely naïve without implantation and seizures. The purpose of these studies was the identification and validation of parameters suitable for an objective severity assessment in rat epilepsy models. Therefore, several behavioral and biochemical parameters were investigated at different time points. To compare these three different models the most interesting period was the chronic phase with generalized seizures. Also, in this phase of the studies most parameters were assessed. For a more detailed description of the models and the studies please see the original published data [1-5].

data(episet_full)

Format

A 943 x 72 data frame with rat epilepsy data.

Details

1. Di Liberto, V., R. M. van Dijk, M. Brendel, A. M. Waldron, C. Möller, I. Koska, I. Seiffert, F. Gualtieri, F. J. Gildehaus, B. von Ungern-Sternberg, M. Lindner, S. Ziegler, R. Palme, R. Hellweg, P. Gass, P. Bartenstein, and H. Potschka. 2018. Imaging correlates of behavioral impairments: An experimental PET study in the rat pilocarpine epilepsy model. Neurobiology of disease 118: 9-21.

2. Koska, I., R. M. van Dijk, I. Seiffert, V. Di Liberto, C. Möller, R. Palme, R. Hellweg, and H. Potschka. 2019. Toward evidence-based severity assessment in rat models with repeated seizures: II. Chemical post-status epilepticus model. Epilepsia 60: 2114-2127.

3. Möller, C., F. Wolf, R. M. van Dijk, V. Di Liberto, V. Russmann, M. Keck, R. Palme, R. Hellweg, P. Gass, C. Otzdorff, and H. Potschka. 2018. Toward evidence-based severity assessment in rat models with repeated seizures: I. Electrical kindling. Epilepsia 59: 765-777.

4. Seiffert, I., R. M. van Dijk, I. Koska, V. Di Liberto, C. Möller, R. Palme, R. Hellweg, and H. Potschka. 2019. Toward evidence-based severity assessment in rat models with repeated seizures: III. Electrical post-status epilepticus model. Epilepsia 60: 1539-1551.

5. van Dijk, R. M., V. Di Liberto, M. Brendel, A. M. Waldron, C. Möller, F. J. Gildehaus, B. von Ungern-Sternberg, M. Lindner, S. Ziegler, R. Hellweg, P. Gass, P. Bartenstein, and H. Potschka. 2018. Imaging biomarkers of behavioral impairments: A pilot micro-positron emission tomographic study in a rat electrical post-status epilepticus model. Epilepsia 59: 2194-2205.

Examples

data(episet_full)
head(episet_full)

#>                   shortname for2591_id species    sex         strain animal_id
#> 1 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 57
#> 2 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 59
#> 3 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 64
#> 4 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 76
#> 5 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 85
#> 6 PostSE_chemical_behaviour        P11     rat female Sprague Dawley     Ko 87
#>   treatment subproject         groups  mod repeats bwc rgs tint_mouse
#> 1     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#> 2     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#> 3     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#> 4     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#> 5     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#> 6     naive        beh chem_beh_naive chem      -1  NA  NA         NA
#>   nesting_rat burrowing_rat openfield_rat cort_plasma cort_fec cort_fur_hair
#> 1          NA        1875.8            NA          NA   3.4797            NA
#> 2          NA        1778.2            NA          NA   1.9845            NA
#> 3          NA        1965.7            NA          NA   2.7930            NA
#> 4          NA        1880.8            NA          NA   3.8000            NA
#> 5          NA        1865.5            NA          NA   3.9837            NA
#> 6          NA        1267.2            NA          NA   3.3264            NA
#>   food social_interaction Sacc_pref OF_rearing_frequency Of_immobile_duration
#> 1   NA                 NA        NA                   NA                   NA
#> 2   NA                 NA        NA                   NA                   NA
#> 3   NA                 NA        NA                   NA                   NA
#> 4   NA                 NA        NA                   NA                   NA
#> 5   NA                 NA        NA                   NA                   NA
#> 6   NA                 NA        NA                   NA                   NA
#>   OF_center_time BWB_WB BWB_entries_in_BB BWB_streching BWB_LT EPM_stretching
#> 1             NA     NA                NA            NA     NA             NA
#> 2             NA     NA                NA            NA     NA             NA
#> 3             NA     NA                NA            NA     NA             NA
#> 4             NA     NA                NA            NA     NA             NA
#> 5             NA     NA                NA            NA     NA             NA
#> 6             NA     NA                NA            NA     NA             NA
#>   EPM_head_dip EPM_closed_arms_duration EPM_open_arms_duration EPM_open1_3
#> 1           NA                       NA                     NA          NA
#> 2           NA                       NA                     NA          NA
#> 3           NA                       NA                     NA          NA
#> 4           NA                       NA                     NA          NA
#> 5           NA                       NA                     NA          NA
#> 6           NA                       NA                     NA          NA
#>   Adrenal_glands Creatincinase Oxytocin BDNF Irwin Seizures_n Seizures_duration
#> 1             NA            NA       NA   NA    NA          0                NA
#> 2             NA            NA       NA   NA    NA          0                NA
#> 3             NA            NA       NA   NA    NA          0                NA
#> 4             NA            NA       NA   NA    NA          0                NA
#> 5             NA            NA       NA   NA    NA          0                NA
#> 6             NA            NA       NA   NA    NA          0                NA
#>   Act_d Act_n HR_d HR_n NN.I_d NN.I_n NNx_d NNx_n pNNx_d pNNx_n RMSSD_d RMSSD_n
#> 1    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#> 2    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#> 3    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#> 4    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#> 5    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#> 6    NA    NA   NA   NA     NA     NA    NA    NA     NA     NA      NA      NA
#>   SDNN_d SDNN_n MPPF_hippocampus_L MPPF_hippocampus_R MPPF_mPFC MPPF_septum
#> 1     NA     NA                 NA                 NA        NA          NA
#> 2     NA     NA                 NA                 NA        NA          NA
#> 3     NA     NA                 NA                 NA        NA          NA
#> 4     NA     NA                 NA                 NA        NA          NA
#> 5     NA     NA                 NA                 NA        NA          NA
#> 6     NA     NA                 NA                 NA        NA          NA
#>   FDG_mPFC FDG_septum FDG_amygdala FDG_striatum FDG_hypothalamus FDG_thalamus
#> 1       NA         NA           NA           NA               NA           NA
#> 2       NA         NA           NA           NA               NA           NA
#> 3       NA         NA           NA           NA               NA           NA
#> 4       NA         NA           NA           NA               NA           NA
#> 5       NA         NA           NA           NA               NA           NA
#> 6       NA         NA           NA           NA               NA           NA
#>   FDG_parietal_cortex FDG_hippocampus Init.ADT Init.SD total.SD Mean.SD
#> 1                  NA              NA       NA      NA       NA      NA
#> 2                  NA              NA       NA      NA       NA      NA
#> 3                  NA              NA       NA      NA       NA      NA
#> 4                  NA              NA       NA      NA       NA      NA
#> 5                  NA              NA       NA      NA       NA      NA
#> 6                  NA              NA       NA      NA       NA      NA